Splits functions out to their own file
