+++ lib/normalizeOptions.js
@@ -1,7 +1,6 @@
+const varPattern = require('./utils/varPattern');
 const { strings, objectsByProperty } = require('./utils/sort');
 const valuesHaveSameUnits = require('./utils/valuesHaveSameUnits');
-
-const varPattern = /var\((--[\w-]+)\)/;
 
 /**
  * Nomalize, collect and merge breakpoint configs.
@@ -97,5 +96,4 @@ function normalizeOptions(options) {
 module.exports = {
   normalizeOptions,
   handleBreakpointConfigs,
-  varPattern,
 };
