+++ lib/normalizeOptions.test.js
@@ -1,4 +1,4 @@
-const { normalizeOptions, handleBreakpointConfigs } = require('./normalize-options');
+const { normalizeOptions, handleBreakpointConfigs } = require('./normalizeOptions');
 
 /**
  * Test expected option values after normalization.
