+++ lib/parseOptions.test.js
@@ -1,4 +1,4 @@
-const { parseOptions } = require('./parse-options');
+const parseOptions = require('./parseOptions');
 
 /**
  * Test CSS option value collection and normalization.
