+++ lib/utils/varPattern.js
@@ -1,0 +1,7 @@
+/**
+ * Pattern to match CSS Custom Properties.
+ * @type {Object}
+ */
+const varPattern = /var\((--[\w-]+)\)/;
+
+module.exports = varPattern;
